ml/skin-cancer-detection/flyte/requirements.txt
flytekit
pandas
opencv-python
numpy
kagglehub
google-cloud-storage
Pillow
matplotlib
seaborn
tensorflow
keras
scikit-learn